[CBRD-22288] fix update xasl's referenced oid list (#1186)
http://jira.cubrid.org/browse/CBRD-22288

OID list of the update's XASL parent was built using the references of the BUILDLIST PROC (it's aptr) which is underneath it. To enable CTEs for Update, CTE procs were placed as aptrs before the last BUILDLIST proc. The fix was to copy the oid list from the last aptr.