[CBRD-22335] Accept unique online indexes. (#1216)
http://jira.cubrid.org/browse/CBRD-22335

Accept unique indexes for online index loading. It has the same behavior as the normal index loading, with an additional check at the end of the loading; after acquiring the strong lock on the table, the index builder will check if the index has the number of keys in the index equal to the number of objects in the index. If so, the uniqueness in correct. Otherwise, it will throw a unique constraint violation.