[CBRD-22397] Skip online index loading for shared constraints (#1235)
http://jira.cubrid.org/browse/CBRD-22397

The normal behavior for shared constraints, specifically for indexes, is that they copy the BTID of the original constraint and they do not start another load process for the new index. Online indexes now follow the same behavior.

The issue resided in the fact that the new index will start another load process that will use the first index's BTID. Therefore, it will begin inserting objects into the BTREE, which was already loaded, and will fail asserts for already inserted objects.