[CBRD-22403] Core dumped in pt_uncorr_post (#1249)
http://jira.cubrid.org/browse/CBRD-22403

Create as subquery creates underneath an insert subquery. Since there is a false where in the non_recursive part of the cte, the non_recursive part is transformed to a false_subquery. This transforms the non_recursive part into a PT_VALUE, making the assert test fail.
It is however alright to just change the assertion since the non_recursive_part will not get its xasl built. Execution does recognize non_recursive_part null as being a false where.