[CBRD-22420] filtered index may not properly end but client misunders… (#1265)
http://jira.cubrid.org/browse/CBRD-22420

It's a bad defect. To build a filtered index may return a bad error code. SCAN_END exits the loop and returned ret which was set by a previous evaluation of filter predicate. In our case, it was V_UNKNOWN.
Server-side index build failed but returned NO_ERROR to client, since no error was set.
The class lock remained as IX and the next online index build hits assertion.