[CBRD-22447] Fix key_len_in_page set for insert_helper (#1273)
http://jira.cubrid.org/browse/CBRD-22447

For the case of insert with DELETE_FLAG in place, we need to check if we have enough space in the leaf page to insert the new object/key. However, this was badly implemented, as the insert_helper did not hold the information of the key's length, and therefore it would always try to insert the key even if there was not enough space. This should fix it.