[CBRD-22502] Fix ovfl keys for a delete which turns into an insert (#1313)
http://jira.cubrid.org/browse/CBRD-22502

In the case of deleting an object which was not yet inserted into the index, we would try to insert it in place if we have enough space available. However, the case of overflow keys was not handled, such as it would try to insert a new key but the btid could not handle it.