[CBRD-22477] - Add code to handle unique indexes for unloaddb (#1365)
http://jira.cubrid.org/browse/CBRD-22477

This PR covers the unique indexes definitions during online index loading. For unique indexes, if there are in the process of being loaded with an online operation, for the unload we cannot tell whether the index is correct or not, as it may not finish. Therefore, we move the definitions to the indexes file resulted from the unload operation, so that DBAs can decide how to load them afterwards.