[CBRD-22624] Fix shared constraints with foreign keys (#1391)
http://jira.cubrid.org/browse/CBRD-22624

For unique indexes and foreign keys, the class constraints would have been decached and we would have lost the shared_cons_name value which was needed to skip the loading. I have added a workaround to tackle this issue by not allowing the online index creation since the btree is already loaded from the shared constraint. This way, we handle it like a normal index which translates into setting the btid of the shared constraint to the current one.