[CBRD-22623] add autowrapping for various json functions (#1421)
http://jira.cubrid.org/browse/CBRD-22623

Autowrapping is performed by json functions with side-effects in case json path last token is an array_index but path's parent is resolved to a scalar when applied to the doc.

Autowrapping works the following way:

    If after the operation performed, the autowrapped scalar becomes an array of n>=2 elements, the wrapping as an array remains: json_insert ({"a":1}', '$.a[1]', 1) = {"a":[1,1]}
    If after the operation performed, the autowrapped scalar becomes a singleton array, the wrapping is reverted with the scalar becoming the singleton element of the array. The operation is performed as if the last token were cut out:
    json_insert ('{"a":1}', '$.a[0]', 2) = json_insert ('{"a":1}', '$.a', 2) = {"a":1}
    json_set ('{"a":1}', '$.a[0]', 2) = json_set ('{"a":1}', '$.a', 2) = {"a":2}

Please note that we accept autowrapping only of the first parent of the pointed path. We diverge from other implementations that do infinite autowrapping (Mysql:

json_set ('{"a":1}', '$.a([0])*', 2) = {"a":2} -- with '*' meaning Kleene operator.