[CBRD-22706] Multi-threaded online index builder (#1431)
http://jira.cubrid.org/browse/CBRD-22706

This is the first implementation of the multi-threaded online index builder. The model of work resembles a thread pool which is managed by the index loader. The index loader scans the heap and fetches an object. The index builder then creates the corresponding key and pushes a task to insert it into the index. The thread pool fetches tasks from the queue and begins inserting into the index. Whenever any thread sets an error, the others will stop their execution when they reach the control code and the loading will then be aborted.