[CBRD-22754] Handle correct worker pool destroy in case of errors (#1466)
http://jira.cubrid.org/browse/CBRD-22754

There was an issue regarding early out errors during online index loading. If there was an error that was not related to the process of loading (i.e.: key generation), then an error was thrown and the process was exited. This meant that the worker pool was not properly destroyed. Now we destroy the pool at any time before exiting the loading process.