[CBRD-22813] Index Builder does not use the global HEAPID when generating compressed strings (#1500)
http://jira.cubrid.org/browse/CBRD-22813

Here, we allocate the new key using the global heap id. However, when we want to get the disk size of the newly created value, if the string compression is enabled it will allocate the compressed string using the heap id of the calling thread. This will fail during the clearing of the value since the clearing is done using the global heap id. This PR should fix it by calling the disk size function when we are using the global heap id.