[CBRD-22869] Implement the --error-file-control argument on the server (#1527)
http://jira.cubrid.org/browse/CBRD-22869

The --error-file-control argument was not implemented on the server. When the argument is provided for loading a database we parse the file and collect errors that need to be filtered by the error handler. We create vector with these errors which is sent to the server and the load workers will treat the cases where an error can be ignored by looking up in the filtered errors vector.