[CBRD-22867] Fix the periodic-commit argument (#1532)
http://jira.cubrid.org/browse/CBRD-22867

In case the periodic commit argument is set to 0, the loading will fail and will crash the server. Since we set the batch size to be equal to the periodic commit, batches will have a size of 0 when the periodic commit is not provided. Therefore, we add a default value to the periodic commit argument in case it is not provided during the loading.