[CBRD-22750] Implement the verbose argument (#1563)
http://jira.cubrid.org/browse/CBRD-22750

Implements the --verbose agument for the new loaddb. It is similar to the error handling implementation. The worker threads append the verbose messages to a string which is fetched on the client along with the statistics. When it arrives at the client, they are printed to stdout and the log file.