[CBRD-23013] Fix error handling in case of string conversion scenarios (#1624)
http://jira.cubrid.org/browse/CBRD-23013

With the new approach, string conversion errors were not correctly handled on SA_MODE. The lexer should throw a fatal error in case it finds such a conversion error, however, on the current SA_MODE, the lexer would just set the session as failed, which was not entirely correct. Now, on SA_MODE we will detect the conversion errors at the time they are encountered by removing the override of LexerError and letting the old implementation handle the error scenario.