[CBRD-22978] Locator multi insert function (#1631)
http://jira.cubrid.org/browse/CBRD-22978

This function should be specialized into inserting an array of records into the same heap file. It works as it follows:

Records passed as parameters for this function should have already been partition pruned before this. Pruning is handled by the insert function.
Insert any records that are not normal ones (REC_HOME) using the old implementation of locator_insert_force.
The normal records are then collected until we manage to fill a whole heap page. (TODO: Needs to be checked).
When the size limit is hit, the insert is issued and we get a new page in which all of the records should have been inserted.

TODO:

Create the heap chain using the pages and append the chain to the heap.
Check for leaks.
Check if the inserts are actually done in the same page!!.