[CBRD-23026] Heap append pages (#1643)
http://jira.cubrid.org/browse/CBRD-23026

This function is able to append a set of pages to a given heap by creating a HEAP_CHAIN using those pages. Then we append the heap chain to the heap by fixing the heap_header_page, followed by fixing the last_page of the heap. Then we fix the first page of the heap chain and we create the new links.