[CBRD-23037] fix heartbeat cluster shutdown deadlock (#1652)
http://jira.cubrid.org/browse/CBRD-23037

heartbeat shutdown thread and heartbeat net poll thread shared same mutex. Therefore when shutdown event is fired the mutex is acquired but the poll thread won't exit since it is blocked on the same mutex.
The deadlock was caused by the fact that heartbeat udp server is stopped under acquired mutex, the fix is to stop the server under no mutex after the cluster is stopped.