[CBRD-22164] fix vacuum_data_load_and_recover after copydb (#1747)
http://jira.cubrid.org/browse/CBRD-22164

Safe-guard failed after copydb because `logpb_last_complete_blockid ()` returns `VACUUM_NULL_LOG_BLOCKID`. Fix the case by skipping `vacuum_Data.set_last_blockid ()` in this case.

This is a hack to reset last_blockid of vacuum data on the first start of copied database (last_blockid is inherited from original database). It is not a proper fix and we should probably find a better one :).

The way now copydb works makes it hard to find non-hackish fix for the issue.

Backport of #1131 and #1139