[CBRD-23097] Limit the number of tasks pushed into a worker pool (#1774)
http://jira.cubrid.org/browse/CBRD-23097

This is a new manager which is able to limit how many tasks are currently running from its worker pool. Whenever a thread wants to execute a task but the worker pull has no workers available, it will wait until one of them finishes its execution in order to complete the task. This is done using a conditional variable which unlocks when the number of tasks available is greater than 0. After finishing its execution, tasks_available gets incremented.