[CBRD-23145] Implement a new lock to isolate multiple threads (#1794)
http://jira.cubrid.org/browse/CBRD-23145

This PR aims to add a new type of lock called BU_LOCK which it is needed for the new loaddb. This new lock is very similar to X_LOCK except the fact that it is compatible with itself. This allows multiple threads in parallel to access the same resource, in loaddb scenario meaning the access on the same heap, but it will isolate external accesses, unless they also use BU_LOCK.

The issue with the previous implementation was that all the workers would try to access the heap using X_LOCK, but only the first caller gains the access and blocks the other threads until it commits. This PR should fix the issue and will later be integrated into loaddb.

Regarding the lock conversion table we decided to go with the following rules:

BU +  IS = X   | IS +   BU  = BU        -> We block other readers until we finish the load.
BU +  IX = X   | IX +   BU  = BU        -> We block other writers until we finish the load.
BU +   S = X   |  S +   BU  =  X
BU +   X = X   |  X +   BU  =  X
BU + SIX = X   | SIX +   BU  = BU