[CBRD-23096] Correct multi-insert logging (#1811)
http://jira.cubrid.org/browse/CBRD-23096

With the current approach we didn't have a mechanism to ensure correctness in case of an abort or crash. This PR aims to fix it by adding a postpone operation over the appending to the heap of the pages created during locator_multi_insert. This way, in case of an abort, the pages would not be appended to the heap meaning that the deallocation would work without any issues.