[CBRD-23173] Loaddb must use BU_LOCK (#1839)
http://jira.cubrid.org/browse/CBRD-23173

After the introduction of the new BU_LOCK we had to use it for the new loaddb. Now workers will acquire BU_LOCK on the class at the beginning of their transaction. Also, we disable class locking during heap_insert_logical and we make sure that the worker has acquired BU_LOCK beforehand.