[CBRD-23226] rename oldest/newest mvccid variables (#1867)
http://jira.cubrid.org/browse/CBRD-23226

Rename variables to clarify their meaning. Replace "active" names with "visible". An MVCCID may be inactive (it was committed/aborted), but it is still visible as long as it is seen as active by concurrent transactions. When all concurrent transactions are finished, the MVCCID becomes invisible.

Vacuum for instance has to wait for an MVCCID to become invisible to clean after it. Inactive is not enough.