[CBRD-23260] Lock-free transaction index (#1883)
http://jira.cubrid.org/browse/CBRD-23260

Add lock-free transaction files and lock-free transaction index concept.

The current design requests and stores pointers to transaction entries for each transaction system in each thread. This causes difficulties to managing lock-free transaction.

First issue is that all transaction systems must be initialized before threads are initialized. That means we need to keep a booking of all needed systems in advance. Note that each lock-free list/hash requires its own system. Therefore, we need to initialize and keep track of all transaction systems that will be used.

We'd like to remove this strict dependency and change the design to something more flexible. Therefore, transaction index concept is added and each thread entry will request and store an index. Transaction systems may be added at any time and each requested index will have an entry in each system.

Please ignore the content of lockfree_transaction files. The only relevance of this content at this time is that some concepts will be renamed to resemble the database transactions that we are familiar to. `lockfree::tran::table` will be equivalent to current `LF_TRAN_SYSTEM` and `lockfree::tran::desc` to `LF_TRAN_ENTRY`. The difference is that each lock-free list and hash will have its own table. Template will also be replaced with a virtual interface.

The term system will now refer to the manager of transaction indexes.