[CBRD-23276] Fixed locking check for rollbacking an insert on unique index (#1897)
http://jira.cubrid.org/browse/CBRD-23276

If a worker fails its inserts on an unique index due to various reasons (e.g. unique constraint violation) it will have to rollback its insert by doing a delete operation. However, the delete operation also requires the locking of the object beforehand. Since we removed the instance locking during loaddb I had to also correct the locking check for the rollback of the insert operation on an unique index.