[CBRD-23266] lock-free freelist (#1886)
http://jira.cubrid.org/browse/CBRD-23266

Implement a class for subset functionality of existing LF_FREELIST. lockfree::freelist provides lock-free claim and retire of preallocated resources, but it does not guarantee no thread is accessing retired resources. This guarantee is to be provided by lock-free transaction system.

Freelist is a template class and can be specialized with any type that has a method with the signature `lockfree::freelist::atomic_link_type &get_freelist_link`, with `using atomic_link_type = std::atomic<T *>`. This link is used by freelist internally. Resources can then be claimed or retired one by one, or retired as a list using freelist link.

There are two tests in the unit test, with 4 threads and 64 threads respectively. The 64 threads test fails, I need to investigate it further.