[CBRD-23290] no oldest unvacuumed update when data is empty (#1905)
http://jira.cubrid.org/browse/CBRD-23290

The existing logic of updating oldest unvacuumed in empty data corner case relies on the assumption that if `lockfree_circular_queue::is_empty ()` return false, following consume must succeed. The assumption is wrong, consume does additional checks (e.g. production is in progress) that may fail.

The fix just simplifies to only update oldest unvacuumed when data is not empty. As consequence, the oldest unvacuumed value may lag behind, but keeping an updated value is not critical. It is used for sanity checks that are still possible with a lagging value.