[CBRD-23373] wait for workers to complete loaddb (#1952)
http://jira.cubrid.org/browse/CBRD-23373

if load session completes before workers, its lock is released and workers are left unprotected. load session complete needs to wait for workers.

added a counter to track active tasks; it is incremented when batch is loaded and decremented when batch finished is notified (transactions are committed/aborted before this, so it is a safe point to decrement the counter). only counter decrements are protected by mutex, increments are atomic (increments are done on same thread that will wait for completion).

removed waiting for completion on interrupt request; load batch request will wait instead.