[CBRD-22741] remove unstable data flush safe-guard (#2018)
http://jira.cubrid.org/browse/CBRD-22741

when flush daemon thread finds no victim candidates, we do many safe-guard checks to see if it there was a fault in searching victim candidates or if it is normal.

one condition is that if list has count_vict_cand greater than zero, then it was expected to find a victim in the list. however, between searching the list and checking safe-guard, new candidates may appear. the condition is not stable and sometimes provides false alarms. fix by removing.