[CBRD-23444] suppress -Wformat-security (#2067)
http://jira.cubrid.org/browse/CBRD-23444

(Warning: "format not a string literal and no format arguments")
Disable, since we could not find a nice solution:
snprintf calls can not be fixed by gcc's __attribute__(__format__(__snprintf__(....))) snprintfs are not supported.
printf (msgcat, args...) are difficult to fix.
Switching to a formating library (e.g. fmt) instead of raw printfs is a possiblity to fix such problems.