[CBRD-23546] wait pending request to close connection (#2134)
http://jira.cubrid.org/browse/CBRD-23546

Connection may be closed between pushing a request from client to worker pool and actually running it. Reusing connection resets the session_p field of CSS_CONN_ENTRY. If the thread that runs this lost request is preempted right when connection is reused, it is possible to generate reported crash.

Fix by tracking pending requests (pushed but not yet executed) and waiting for all of them to be accounted before closing connection.