[CBRD-23340] Fix bad line report during error handle (#2117)
http://jira.cubrid.org/browse/CBRD-23340

The old implementation of loaddb would take into account different cases of error reporting and would use a manual adjustment of the line numbering in order to display the correct line. The issue that led to the creation of this adjustment is that the scanner will auto increment the line as it ends it, and error that would come from the server insert should have been reported on the previous line.

This PR aims to somehow imitate the behavior in an automatic way. Instead of manually adjusting the lineno, we just memorize the starting line of the current record, and if we encounter any type of error in parsing or inserting, we report the error as the starting line of the record.