[CBRD-23600] Unexpected rounding up a numeric value when the case expressions are used recursively (#2180)
http://jira.cubrid.org/browse/CBRD-23600

When the CASE expressions are used recursively and values specified in the THEN clause are numeric and integer type of order repeatedly, one of the numeric values is rounded up unexpectedly. It causes because the arg1 pointer variable for PT_NODE is incorrectly re-used at the part of checking PT_NODE of the next case clause. This commit fixes to keep the arg1 from being re-used incorrectly.