Set TDE along with f_init and not reset for temp
Now, we see the f_init()ed page is allocated properly, which is
reasonable becuase if the page tpye is unknown, it is regarded as
dellcated in pgbuf_fix(). So we set tde algorithm only after
f_init() is applied.

With it, in the case of temp file, we don't reset the tde information
when retiring temp file and reset it in file_alloc(). It was all reset in
retiring before because fixing page to reset the tde algorithm in
page_alloc would be expensive, but now to do f_init(), it is already
fixed and cause nearly no cost.