[LETS-24] Handle SA_MODE server_type & communication channels (#2572)
http://jira.cubrid.org/browse/LETS-24

libcubridsa compile can fail because communication channel function definitions are missing. Fix by changing how server_type & communication channels are handled in SA_MODE.

Server type is currently force to be always TRANSACTION in SA_MODE. Currently, no connections to a page server is allowed, but this may be changed in the future.

The communication channels are handled similarly as in HA redesign. Base communication channel is included in SA_MODE too. Server communication channel is not included in SA_MODE, but since the code that connects to the page server is disabled, the compile works. If SA_MODE server will connect to the page server, server communication channel will have to be included in SA_MODE too.