[LETS-20] log_recovery_redo refactoring to prepare for parallelization (#2577)
Re-arrange code in (and around) `log_recovery.log_recovery_redo` in preparation for introducing a parallelization mechanism for applying logs:
- transform `LOG_READ_*` macros into inline functions
- introduce `log_reader` class which manages access and usage of a `log_lsa` and `log_page` pointer whereas before they were handled "inlined" in `log_recovery_redo` function; integrate this class in `log_recovery_redo` and downstream functions
- re-arrange code that extracts - potentially archived - `undo` and `redo` data into functions `log_rv_get_unzip_log_data`, `log_rv_get_unzip_and_diff_redo_log_data` and use in `log_rv_redo_record` function
- `log_recovery_redo`: moved c-style function level local variables to innermost scopes - this aids in further  refactoring

This is a preliminary change for the next step: refactor the big `switch` statement in `log_recovery_redo` by separating each case and delegating as much as possible to the - what already exists - `log_rv_redo_record` function. This next step is to be written with the explicit goal of allowing natural further delegation of certain steps to a parallelized mechanism of applying redo log entries.

Other:
- the `log_reader` class can probably be integrated in other `log_recovery` functions as well