[LETS-8] Disable writing (and log creation) on the page server (#2586)
http://jira.cubrid.org/browse/LETS-8

Queries and log creation should be disabled on the page server. Other tools that operate changes on the database like addvoldb, compactdb, loaddb, optimizedb, synccoldb, are also disabled. New error messages will appear when trying to run changes on the page server, or when the server is appending a log record.