[LETS-33] parallel log recovery redo (#2619)
http://jira.cubrid.org/browse/LETS-33

This is the concluding change-set that introduces the mechanism to apply log recovery redo in parallel.

The mechanism itself is implemented in `log_recovery_redo_parallel` files and is made up of a set of classes that together implement an M-to-N prod-consumer async mechanism. On the prod side, there is only one thread - the main thread supplying redo jobs (`redo_job_impl`). On the consumer side there is a number of tasks (`redo_task`) that consume the redo jobs. The actual number of tasks is defined defined by the system parameter `PRM_ID_RECOVERY_PARALLEL_COUNT` (default 8). If the value of this parameter is 0 the parallel async mechanism is not used at all - instead, the logic works completely sync/sequential as it was before this change.
The tasks consume (ie: apply) log redo records abiding by the following rules:
- within one page (defined by a VPID) log redo records are applied in the same order they appear in the log;
- at a certain moment, there is only one task that applies one log redo record on a page

To use this infrastructure, an encapsulating facade class - `redo_parallel` - is used.

Integration:
- function `log_recovery_redo`: depending on build type and system parameter, instantiate `redo_parallel` with the number of workers; this creates the async threads (tasks)
- function `log_rv_redo_record_sync_or_dispatch_async`: if the log redo record entry has a non-null VPID, a task is created and dispatched which applies the log redo record on that page;
- a number of exceptions are in place which are always applied synchronously:
  - all null VPID log redo records
  - all types specified by the function `log_rv_need_sync_redo`
- `RVDK_UNRESERVE_SECTORS`: To apply RVDK_UNRESERVE_SECTORS, one must first wait for all changes in this sector to be redone. Otherwise, asynchronous jobs skip redoing changes in this sector's pages because they are seen as deallocated. When sector is reserved again, redo is resumed in the sector's pages, but the pages are not in a consistent state. The current workaround is to wait for all changes to be finished, including changes in the unreserved sector.
- after all log redo records have been processed (or dispatched async), the logic waits for the async ones to be applied before terminating the redo phase (end of `log_recovery_redo` function)