[LETS-31] Add a class for all checkpoint information (#2678)
Checkpoint information is saved to disk as end checkpoint log record. The information must be moved to another file. A step to improve handling checkpoint is to create a class that holds all the required information and all required operations.

As the first part of the changes, a new class has been added along with the its own serialization method.