[LETS-32] Add new log volume to store meta log (#2700)
http://jira.cubrid.org/browse/LETS-32

Add a new volume for the meta log (checkpoint info). In this patch, the meta log volume is empty. The actual writing of meta log (and checkpoint info) will be done in a subsequent task.

The new file is named dbname_lgmeta, similarly to other log files and it is stored in the configured log directory. The file is created during createdb and deleted during deletedb. It is copied during copydb and is backed-up/restored.

The meta log is loaded from file when log is initialized. Not finding the meta log file is acceptable when emergency initialization is called (e.g. deletedb).