[LETS-115] Reset nxio_lsa to append_lsa during boot/recovery (#2795)
http://jira.cubrid.org/browse/LETS-115

Pages are requested during boot/recovery, before any log being flushed to disk and thus, before setting any value to nxio_lsa. These requests are sent with a null target replication LSA and trigger an assert on the page server.

The pages need to be sent with a valid target replication LSA. E.g. if page server replication is a couple of seconds behind, and if the transaction server restarts, it needs to wait until all the changes in the requested pages are replicated.

nxio_lsa is the LSA after the last log record written to disk. During boot/recovery, append LSA is reset to the same value with the
LOG_RESET_APPEND_LSA function. Same function is used to also set the value of nxio_lsa.

Some pages (e.g. data volume headers) are requested before the log module is initialized. The transaction server has no idea at this point what target LSA to set for its data page requests. For now, the wait for replication is disabled when target_lsa is null. In the future, the transaction server should set a good target from the start (probably with the help of page server, by obtaining some information from there).