[LETS-95] Only perform consistency check in recovery mode, not in replication mode #2799
http://jira.cubrid.org/browse/LETS-95

When executing the recovery phase of the boot process log_Gl.rcv_phase is moved through a few values and ends up, at the end of the recovery process, with the LOG_RESTARTED value.
Use this flag to only execute the consistency check when in actual recovery, not when in "normal" mode. Thus, the code is not executed for replication on the page server side. Eventually, this piece of instrumentation will be removed altogether.