[LETS-123] Always make sure the log is flushed to disk on the page server (#2805)
http://jira.cubrid.org/browse/LETS-123

Running many operations in a single transaction on the transaction server sometimes hangs. The reason is that the log flush thread on the page server misses a wakeup call and never flushes the last batch of log records. The transaction server does nothing but waits for a confirmation from the page server that never comes.

Two reasons have been discovered for missing the confirmation:

  1. A wakeup call that came while a log flush iteration was still in progress was lost because the requested flag is reset at the end of log flush execution. In the next iteration it earlies out.
  2. Prior list has been serialized to log pages, but was not flushed to disk.

log_flush_execute (the loop function of log flush daemon) has fixes for both problems:

  1. Use compare exchange when testing the requested flag, from true to false. If it returns true, it means there are requests. Any wakeup calls during log_flush_execute will not be missed.
  2. Even if no wakeup requests are registered, on the page server, if there are any unflushed log records in prior list or appended in the log pages, flush is forced.