[LETS-122] Wait for replication to catch-up when running checkpoint on the page server. (#2804)
http://jira.cubrid.org/browse/LETS-122

On the transaction server, the regular order of operating database changes is:

  1. Fix data page
  2. Log change on the page &  make change in the page
  3. Unfix page.

When checkpoint gets its checkpoint LSA, all changes up to checkpoint LSA are already reflected in the pages that will be flushed to disk.

On the page server, the log comes from the transaction server. The log records exist way before the change is reflected into the data page. When checkpoint got its checkpoint LSA, the replication was behind and changes that occurred before checkpoint LSA were missed while pages were being flushed to disk. This caused consistencies caught by checkpoint safe-guards.

Fix by waiting for the replication to catch-up with the checkpoint LSA before proceeding to flushing data pages to disk.