[LETS-64] Initialize log prior receiver only on the page server (#2820)
http://jira.cubrid.org/browse/LETS-64

Log prior receiver automatically starts a thread when it is constructed, ready to take over transferred log prior lists. On the active transaction server this thread does nothing, because the server does not receive any log.

Fix by changing the log_Gl.m_prior_recver member to a unique pointer. The log prior receiver is initialized with init_server_type(), only on the page server, and it is finalized during shutdown before log_final() is called.